unify fetching list of implications
